Task #2 Created a component that implements react-mailchimp-subscribe lib, handles mailchimp url/secret and abstracts a child, which will receive the props from this component (subscribe, status, and message). It also centralizes the role of mailchimp. <Newsletter> changes are in place and working fine. Missing <ContactUs>.
