+++ components/Newsletter/NewsletterForm/index-b.js
@@ -1,0 +1,7 @@
+// const NewsletterForm = ({ status, message, onValidated }) => {
+const NewsletterForm = props => {
+  console.log('props on NewsletterForm: ', props);
+  return <h1>NEWSLETTER-FORM</h1>;
+};
+
+export default NewsletterForm;
