+++ components/Newsletter/NewsletterForm/index.js
@@ -7,44 +7,52 @@ import S from './styles';
 
 const SITE_KEY = process.env.NEXT_PUBLIC_RECAPTCHA_SITE_KEY;
 
-const NewsletterForm = ({ status, message, onValidated }) => {
-  const [error, setError] = useState(null);
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const recaptchaRef = createRef();
+const NewsletterForm = ({ status, message, subscribe }) => {
+  console.log('props on NLF::: ', status, message, subscribe);
+  /////////////////// temp stuff
+  const t = new Date();
+  const tday = t.getDate() > 9 ? t.getDate() : '0' + t.getDate();
+  const thour = t.getHours() > 9 ? t.getHours() : '0' + t.getHours();
+  const tmonth =
+    t.getMonth() + 1 > 9 ? t.getMonth() + 1 : '0' + (t.getMonth() + 1);
+  const tminute = t.getMinutes() > 9 ? t.getMinutes() : '0' + t.getMinutes();
+  const tname = tday + '' + tmonth + '' + thour + '' + tminute;
+  const temail = `tony.kieling+${tname}@gmail.com`;
+  const [name, setName] = useState(tname);
+  const [email, setEmail] = useState(temail);
 
-  const onReCAPTCHAChange = async captchaCode => {
+  const [error, setError] = useState(null);
+  // const [name, setName] = useState('');
+  // const [email, setEmail] = useState('');
+  const recaptchaRef = createRef();
+  const [reCaptchaFail, setReCaptchaFail] = useState(false);
+
+  const onReCAPTCHAChange = () => {
+    console.log('new recaptcha!!!!!!!!!!!!!');
+    recaptchaRef.current.reset();
+  };
+
+  const validateReCaptcha = async reCaptchaToken => {
     // If the reCAPTCHA code is null or undefined indicating that
     // the reCAPTCHA was expired then return early
-    if (!captchaCode) {
+    if (!reCaptchaToken) {
+      setReCaptchaFail(true);
       return;
     }
+
     try {
-      const response = await fetch('/api/register', {
+      const response = await fetch('/api/validateReCaptcha', {
         method: 'POST',
-        body: JSON.stringify({ email, name, captcha: captchaCode }),
+        body: JSON.stringify({ email, name, gReCaptchaToken: reCaptchaToken }),
         headers: {
           'Content-Type': 'application/json',
         },
       });
-      if (response.ok) {
-        // If the response is ok than show the success console.log
-        console.log('Email registered successfully');
-      } else {
-        // Else throw an error with the message returned
-        const error = await response.json();
-        throw new Error(error.message);
-      }
+
+      return response.ok ? true : false;
     } catch (error) {
       console.log(error?.message || 'Something went wrong');
-    } finally {
-      // Reset the reCAPTCHA when the request has failed or succeeded
-      if (recaptchaRef?.current) {
-        recaptchaRef.current.reset();
-      }
-
-      setEmail('');
-      setName('');
+      return false;
     }
   };
 
@@ -53,9 +61,10 @@ const NewsletterForm = ({ status, message, onValidated }) => {
    *
    * @return {{value}|*|boolean|null}
    */
-  const handleFormSubmit = event => {
+  const handleFormSubmit = async event => {
     event.preventDefault();
 
+    setReCaptchaFail(false);
     setError(null);
 
     if (!name) {
@@ -68,17 +77,29 @@ const NewsletterForm = ({ status, message, onValidated }) => {
       return null;
     }
 
-    recaptchaRef.current.execute();
+    // const validateRecaptcha = await recaptchaRef.current.execute();
+    const token = await recaptchaRef.current.executeAsync();
+    console.log('----------going to check recaptcha', token);
+    const confirmValidateRecaptcha = await validateReCaptcha(token);
+    console.log('confirmValidateRecaptcha::: ', confirmValidateRecaptcha);
 
-    const isFormValidated = onValidated({
-      EMAIL: email,
-      MERGE1: name,
-    });
+    if (!confirmValidateRecaptcha) {
+      setReCaptchaFail(true);
+    } else {
+      // const isFormValidated  -- onValidate return nothing, so no need it
+      subscribe({
+        EMAIL: email,
+        MERGE1: name,
+      });
 
-    event.target.reset();
+      // console.log("isFormValidated::: ", isFormValidated)
+      event.target.reset();
+      setEmail('');
+      setName('');
+      setReCaptchaFail(false);
+    }
 
-    // On success return true
-    return name && email && email.indexOf('@') > -1 && isFormValidated;
+    // recaptchaRef.current.reset();
   };
 
   /**
@@ -157,6 +178,13 @@ const NewsletterForm = ({ status, message, onValidated }) => {
           </S.Form>
 
           <S.FormInfo>
+            {/* just in case when reCaptcha issue happens */}
+            {reCaptchaFail && (
+              <S.FormSending>
+                Please, refresh your screen and try again.
+              </S.FormSending>
+            )}
+
             {status === 'sending' && <S.FormSending>Sending...</S.FormSending>}
             {status === 'error' || error ? (
               <S.FormError
